requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
Bottleneck==1.4.2
brotlicffi==1.0.9.2
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
datasets==4.0.0
Deprecated==1.2.13
dill==0.3.8
filelock==3.17.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2==2.2.1
hf-xet==1.1.8
huggingface_hub==0.35.0
idna==3.7
importlib_metadata==8.5.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.15
narwhals==2.5.0
neo4j==6.0.2
networkx==3.4.2
numexpr==2.11.0
numpy==2.2.5
opentelemetry-api==1.30.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
propcache==0.3.1
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.23
pydeck==0.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rpds-py==0.22.3
safetensors==0.5.3
setuptools==72.1.0
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1a0
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.55.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==4.0.2
wheel==0.45.1
wrapt==1.17.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
